Rename dirs to avoid creating lambdas
