+++ web/index.ts
@@ -1,4 +1,4 @@
-import { ParsedRequest, Theme, FileType } from '../api/types';
+import { ParsedRequest, Theme, FileType } from '../api/_lib/types';
 const { H, R, copee } = (window as any);
 let timeout = -1;
